Add ProviderMixin
This is an approach to a DI-like system for web components covered in [this YouTube video](https://youtu.be/x9YDQUJx2uw?t=2104). The general idea is that a parent component registers itself as a provider, and child components can request data from that provider. The specific problem that this aims to solve is preventing chained parameter-passing through multiple levels of components, so that a deep component in the app can have some config or data. By instead using this pattern, only the provider component and the consumer component need to know anything about that config/data.

Just putting this up for consideration right now - we're using this pattern in the FACE app, and it seems like it would be relevant to other places as well. In our use case, the top-level component that gets rendered by the LMS also gets some configs slapped onto it as attributes, which it then registers itself as a provider of, and a child component several layers down requests said config.

If this makes sense to exist in core, I'll add tests, but for now just opening up the discussion.