When an element or one of its parents is set to "display: 'none'", all in-progress animations will be cancelled and any new animations will be queued but never run until the element is made visible again.
This causes two problems:
1. When trying to reset a hidden hierarchical-view to open on its main view, the view won't reset
until the hierachical-view is made visible again and the animation can run.
2. When an animation is in progress and a hierarchical-view's display is set to 'none', the animation
will get cancelled. This animation will then get replayed from the start when the view becomes visible
instead of having finished while hidden.

These issues are solved by not running animations and cancelling any in-progress animations if the
hierachical-view is hidden while the display is set to 'none';

Using an animationcancel event would be ideal instead of tracking animations with the __hideAnimations
property but it cannot be used because it is only supported by Firefox.