Fix issue where the d2l-visible-on-ancestor-target may not have loaded
When the connectedCallback is run, there is no guarantee that the ancestors' shadow
DOM have loaded. If the d2l-visible-on-ancestor target is within an ancestor's
shadow DOM then the look up in __voaInit will fail and the visible-on-ancestor-mixin
will not find a target.

To account for the async loading of ancestors' shadow DOM, this delays until the
next animation frame.