Remove negative width check
The polymer tooltip has a bug where if the left boundary is greater than the right
boundary, the right boundary will be ignored. This behavior is being relied on in
at least one known place but potentially more for correct positioning. This will
be addressed in those places when the migration happens rather than introducing
a bug for backwards compatibility.