Move the target-position styling from being on and internal div to be on the tooltip itself
This change was made because the tooltip can potentially have an offsetParent meaning it ends up off the page.
In situations like these the tooltip would still position and size correctly but the 0 by 0 tooltip itself
would be off the page and trigger horizontal or vertical scroll.

To avoid this, the tooltip itself needs to be positioned to align with its target rather than just moving
the internal positioning div.