Fix defect were transition would get skipped periodically when transitioning from display: none or from auto to 0
This was fixed by adding a double requestAnimationFrame wait. The reason for this is that when a Lit property gets updated, the browser
will not immediately perform a style recalc. The next style recalc will happens after the next requestAnimationFrame. If we try to update
styles again in the first requestAnimationFrame block then the first style change from the property change will be overwritten and never
get displayed by the browser causing the transition to be skipped. In the second requestAnimationFrame we can guarantee that the property
style change has taken effect so it is now safe to make our dependent style update.