Fix positioning defect
In Chrome offsetParent includes positioned nodes in ancestors' shadow DOM. In Firefox offsetParent positions relative to these
but reports the first offsetParent as the first positioned ancestor in the same DOM context. This causes offsetTop and offsetLeft
to be incorrect in Firefox if there is a positioned node in an ancestor's shadow DOM. To fix this, offsetTop and offsetLeft need
to be computed using the getOffsetParent DOM helper rather than using the built-in properties.