Add showErrors parameter to validate
This allows validate to be called with the purpose of only hiding the error message on validate.
If a new error occurs and showErrors is false then the component will not be put in an invalid state.
This allows errors to be shown when requestValidate(true) is called on blur.