Replace form-element showErrors with validationType
ValidationType provides more control over how errors are rendered in a custom form element's sub-tree.
SHOW_NEW_ERRORS causes validation to show any errors that occurred immediately. This can be used during blur or change to notify users that their input was invalid.
UPDATE_EXISTING_ERRORS causes validation to only update an error message if the error message changed to a different one or remove it if the input become valid. This should be used during live validation when a user is in the process of input so we don't show new errors if they are half-way through inputing something like the start of an email address.
SUPPRESS_ERRORS will validate but supress any errors that occur. This can be used in the case of nested custom form elements so that the parent won't show any validation errors if the children that the parent depends on are invalid. This might needed if the parent's error message overlaps with a child's error message.