Fix tryGetLabelText in Safari when used on elements in the Shadow
The labels property is fully supported in Safari but is broken when called on an
element in the shadow DOM. In this case it always returns an empty list even
if the element has an implicit or explicit label.

To fix it, we need to manually check the parent for an implicit label or look
for labels with a for attribute pointing to the label.