Add the MobileTouchResizer
This resizer resizes the secondary panel in mobile drawer mode
when the secondary panel is scrolled up or down. When users
attempt to scroll the secondary panel and it isn't fully open
the panel will be opened. When users attempt to scroll the
secondary panel down and its scrolled to the top it will
cause the panel to collapse. This also applies to small scrolls
that act like swipe gestures.