Add icon to d2l-alert
While developing the @brightspace-ui-labs/media-player,
we wanted to display an alert to the user. Some documentation
(http://design.d2l/components/alerts/) shows that d2l-alerts
can support images. Upon slotting an <img>, I realized that it
was moving things around within d2l-alert.

Instead of adding support for images which may come in all shapes
and sizes, I decided to add support for the d2l-icons that we have.