Close drop-dropdown menu on link select
Selecting a menu item did not previously close menus,
because they were expected to navigate to another page.
If the target is a new tab, the original page would remain
intact, with the drop-down menu still open.

This change will close drop-down menus whenever a menu
item link is selected.