Fix secondary panel height
If dropdowns or tooltips open in the secondary panel they may get cut off. This
happens because the aside inside the secondary panel only takes up the height of its
content, not the height of its panel.

This is fixed by forcing the aside to always be the height of the panel