fix: demo snippet styling
This allows elements that want to use scroll wrapper and overflow to display
correctly. It will hide the overflow and allow scrolling with arrow buttons.
This also makes the code view always be the same width as the demo snippet when
it is part of the snippet. It will still have the maximum width by itself.