Fix issue with Prim/Sec footer padding on print
Closes: https://desire2learn.atlassian.net/browse/BORG-5377

Normally not a noticable issue because the footer is below all content
when printing. However, when using CSS to prevent page breaking in the
middle of content, those content elements get pushed further down the
page but the footer does not, resulting in the footer overlapping and
the padding "erasing" some of the content.